This commit fixes the following:
The network type was not set to correct value in configure_network_settings module. This was fixed.

The settings for disable automatic to False has been added to migration and rollback both.

Test for configuring params during migration and rollback are added.

Also some checks on which values can be passed for a cni during migration and rollback have been added.