Some changes in run_command to remove shell=true. Right practices to not hand-writing JSON.
Fixed some issues in disable auto migration during rollback which was a result of doc bug where networkType is set
two times at step 4 and step 7 in https://docs.openshift.com/container-platform/4.16/networking/ovn_kubernetes_network_provider/rollback-to-openshift-sdn.html
Also added a check as per step 5 in the rollback.

While doing check for mcp completion we check if the chain of condtions does not return any error and in case of no error we exit the loop bu returning True.